Move json_reform functions from policy.py to calculate.py
