+++ taxcalc/calculate.py
@@ -11,6 +11,9 @@ Tax-Calculator federal tax Calculator class.
 # pylint: disable=no-value-for-parameter,protected-access
 
 
+import os
+import json
+import re
 import copy
 import numpy as np
 from .utils import *
@@ -371,3 +374,57 @@ class Calculator(object):
         calc = Calculator(policy=clp, records=recs, sync_years=False,
                           behavior=behv, growth=grow, consumption=cons)
         return calc
+
+    @staticmethod
+    def read_json_reform_file(reform_filename):
+        """
+        Read reform file, strip //-comments, and return dict based on JSON.
+        The reform file is JSON with string policy-parameter primary keys and
+           string years as secondary keys.  See tests/test_policy.py for an
+           extended example of a commented JSON reform file that can be read
+           by this method.
+        Returned dictionary has integer years as primary keys and
+           string policy-parameters as secondary keys.
+        The returned dictionary is suitable as the argument to the
+           implement_reform(reform_dict) method (see below).
+        """
+        if os.path.isfile(reform_filename):
+            txt = open(reform_filename, 'r').read()
+            return Calculator.read_json_reform_text(txt)
+        else:
+            msg = 'reform file {} could not be found'
+            raise ValueError(msg.format(reform_filename))
+
+    @staticmethod
+    def read_json_reform_text(text_string):
+        """
+        Strip //-comments from text_string and return dict based on JSON.
+        The reform text is JSON with string policy-parameter primary keys and
+           string years as secondary keys.  See tests/test_policy.py for an
+           extended example of a commented JSON reform text that can be read
+           by this method.
+        Returned dictionary has integer years as primary keys and
+           string policy-parameters as secondary keys.
+        The returned dictionary is suitable as the argument to the
+           implement_reform(reform_dict) method (see below).
+        """
+        # strip out //-comments without changing line numbers
+        json_without_comments = re.sub('//.*', ' ', text_string)
+        # convert JSON text into a dictionary with year skeys as strings
+        try:
+            reform_dict_raw = json.loads(json_without_comments)
+        except ValueError as valerr:
+            msg = 'Policy reform text below contains invalid JSON:\n'
+            msg += str(valerr) + '\n'
+            msg += 'Above location of the first error may be approximate.\n'
+            msg += 'The invalid JSON reform text is between the lines:\n'
+            bline = 'XX----.----1----.----2----.----3----.----4'
+            bline += '----.----5----.----6----.----7'
+            msg += bline + '\n'
+            linenum = 0
+            for line in json_without_comments.split('\n'):
+                linenum += 1
+                msg += '{:02d}{}'.format(linenum, line) + '\n'
+            msg += bline + '\n'
+            raise ValueError(msg)
+        return Policy.convert_reform_dictionary(reform_dict_raw)
