+++ taxcalc/tests/test_policy.py
@@ -500,101 +500,6 @@ def test_parameters_get_default_start_year():
     assert meta_kt_c_age['value'] == [24]
 
 
-REFORM_CONTENTS = """
-// Example of a reform file suitable for Policy read_json_reform_file function.
-// This JSON file can contain any number of trailing //-style comments, which
-// will be removed before the contents are converted from JSON to a dictionary.
-// The primary keys are policy parameters and secondary keys are years.
-// Both the primary and secondary key values must be enclosed in quotes (").
-// Boolean variables are specified as true or false (no quotes; all lowercase).
-{
-    "_AMT_brk1": // top of first AMT tax bracket
-    {"2015": [200000],
-     "2017": [300000]
-    },
-    "_EITC_c": // maximum EITC amount by number of qualifying kids (0,1,2,3+)
-    {"2016": [[ 900, 5000,  8000,  9000]],
-     "2019": [[1200, 7000, 10000, 12000]]
-    },
-    "_II_em": // personal exemption amount (see indexing changes below)
-    {"2016": [6000],
-     "2018": [7500],
-     "2020": [9000]
-    },
-    "_II_em_cpi": // personal exemption amount indexing status
-    {"2016": false, // values in future years are same as this year value
-     "2018": true   // values in future years indexed with this year as base
-    },
-    "_SS_Earnings_c": // social security (OASDI) maximum taxable earnings
-    {"2016": [300000],
-     "2018": [500000],
-     "2020": [700000]
-    },
-    "_AMT_em_cpi": // AMT exemption amount indexing status
-    {"2017": false, // values in future years are same as this year value
-     "2020": true   // values in future years indexed with this year as base
-    }
-}
-"""
-
-
-@pytest.yield_fixture
-def reform_file():
-    """
-    Temporary reform file for Policy read_json_reform_file function.
-    """
-    rfile = tempfile.NamedTemporaryFile(mode='a', delete=False)
-    rfile.write(REFORM_CONTENTS)
-    rfile.close()
-    # must close and then yield for Windows platform
-    yield rfile
-    if os.path.isfile(rfile.name):
-        try:
-            os.remove(rfile.name)
-        except OSError:
-            pass  # sometimes we can't remove a generated temporary file
-
-
-@pytest.mark.parametrize("set_year", [False, True])
-def test_read_json_reform_file_and_implement_reform(reform_file, set_year):
-    """
-    Test reading and translation of reform file into a reform dictionary
-    that is then used to call implement_reform method.
-    NOTE: implement_reform called when policy.current_year == policy.start_year
-    """
-    reform = Policy.read_json_reform_file(reform_file.name)
-    policy = Policy()
-    if set_year:
-        policy.set_year(2015)
-    policy.implement_reform(reform)
-    syr = policy.start_year
-    amt_brk1 = policy._AMT_brk1
-    assert amt_brk1[2015 - syr] == 200000
-    assert amt_brk1[2016 - syr] > 200000
-    assert amt_brk1[2017 - syr] == 300000
-    assert amt_brk1[2018 - syr] > 300000
-    ii_em = policy._II_em
-    assert ii_em[2016 - syr] == 6000
-    assert ii_em[2017 - syr] == 6000
-    assert ii_em[2018 - syr] == 7500
-    assert ii_em[2019 - syr] > 7500
-    assert ii_em[2020 - syr] == 9000
-    assert ii_em[2021 - syr] > 9000
-    amt_em = policy._AMT_em
-    assert amt_em[2016 - syr, 0] > amt_em[2015 - syr, 0]
-    assert amt_em[2017 - syr, 0] > amt_em[2016 - syr, 0]
-    assert amt_em[2018 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2019 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2020 - syr, 0] == amt_em[2017 - syr, 0]
-    assert amt_em[2021 - syr, 0] > amt_em[2020 - syr, 0]
-    assert amt_em[2022 - syr, 0] > amt_em[2021 - syr, 0]
-    add4aged = policy._ID_Medical_frt_add4aged
-    assert add4aged[2015 - syr] == -0.025
-    assert add4aged[2016 - syr] == -0.025
-    assert add4aged[2017 - syr] == 0.0
-    assert add4aged[2022 - syr] == 0.0
-
-
 def test_pop_the_cap_reform():
     """
     Test eliminating the maximum taxable earnings (MTE)
@@ -661,25 +566,6 @@ def test_misspecified_reforms():
     # 2016 key:value pair in reform2 (2016:{'_II_em...}) overwrites and
     # replaces the first 2016 key:value pair in reform2 (2016:{'_SS_E...})
     assert not reform1 == reform2
-
-
-@pytest.yield_fixture
-def badreformfile():
-    # specify JSON text for policy reform
-    txt = """{ // example of incorrect JSON because 'x' must be "x"
-               'x': {"2014": [4000]}
-             }"""
-    f = tempfile.NamedTemporaryFile(mode='a', delete=False)
-    f.write(txt + '\n')
-    f.close()
-    # Must close and then yield for Windows platform
-    yield f
-    os.remove(f.name)
-
-
-def test_read_bad_json_reform_file(badreformfile):
-    with pytest.raises(ValueError):
-        Policy.read_json_reform_file(badreformfile.name)
 
 
 def test_convert_reform_dictionary():
