+++ taxcalc/tests/test_reforms.py
@@ -11,7 +11,7 @@ import glob
 import re
 import json
 import pytest
-from taxcalc import Policy  # pylint: disable=import-error
+from taxcalc import Calculator, Policy  # pylint: disable=import-error
 
 
 @pytest.fixture(scope='session')
@@ -35,7 +35,7 @@ def test_reforms(reforms_path):  # pylint: disable=redefined-outer-name
         with open(jrf) as jrfile:
             jrf_text = jrfile.read()
         # check that jrf_text can be implemented as a Policy reform
-        jrf_dict = Policy.read_json_reform_text(jrf_text)
+        jrf_dict = Calculator.read_json_reform_text(jrf_text)
         policy = Policy()
         policy.implement_reform(jrf_dict)
         # identify policy parameters included in jrf after removing //-comments
